requirements.txt
web3==7.0.06b
pandas==2.0.3
sqlalchemy==1.4.27
python-dotenv==0.20.0
psycopg2==2.9.9
retry==0.9.2